feat(*): fix subsequent dragging bug, refactor to pass methods from cicero-ui to markdown-editor
